--- a/src/util.js
+++ b/src/util.js
@@ -209,19 +209,16 @@ const cursorFormatter = {
 };
 
 function removeFixings({ prefix = "", postfix = "", value = "" } = {}) {
-  const start = value.toLowerCase().startsWith(prefix.toLowerCase())
-    ? prefix.length
-    : 0;
+  if (prefix === "" && postfix === "") return value;
+
+  const valueLower = value.toLowerCase();
+
+  const start = valueLower.startsWith(prefix.toLowerCase()) ? prefix.length : 0;
   const end =
     value.length -
-    (value.toLowerCase().endsWith(postfix.toLowerCase()) ? postfix.length : 0);
+    (valueLower.endsWith(postfix.toLowerCase()) ? postfix.length : 0);
 
-  let formatted = "";
-  for (let i = start; i < end; i++) {
-    formatted += value[i];
-  }
-
-  return formatted;
+  return value.slice(start, end);
 }
 
 function addPadding({ padding = {}, value = "" } = {}) {