@anatolzak 2025-08-09
hey @tywalch!
I hope you are well! I know we have another significant PR in progress, but do you think we could prioritize this one? It significantly improves Electro's performance, which currently has over 100ms of overhead in certain situations, which is not ideal. This PR reduces it to ~15 ms. The changes in this PR are straightforward but would probably provide substantial benefits to most Electro users.
Thanks so much 🙏